refactor: Split functions into files
