--- src/js/utils.js
@@ -1,3 +1,0 @@
-export const sayHello = (text) => {
-  return text;
-};
